% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_selections.R
\name{add_selections}
\alias{add_selections}
\title{add_selections function}
\usage{
add_selections(unit_table)
}
\arguments{
\item{unit_table:}{A data.frame with the columns specified in the description.}
}
\value{
A new unit table with a sel column.
}
\description{
Takes a unit table a adds a new column called "sel", indicating the number 
of the selected note in each song. (The k^th note would have sel k). Input unit table must
have columns: 1) "sound.files": name of the .wav files,
2) "sel": number of the selections, 
3) "start": start time of selections, 
4) "end": end time of selections
}
\examples{
unit_table = tibble::tibble(sound.files = c( rep("JS001.wav",3), rep("JS002.wav",3)),
start = seq(from = 1, to = 3.5, by = 0.5),
end = seq(from = 1, to = 3.5, by = 0.5) + 0.5)
}
