% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transitions.R
\name{add_transitions}
\alias{add_transitions}
\title{add_transitions function}
\usage{
add_transitions(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{

}
\description{
Takes in a unit table of notes and adds a new column
indicating the next adjacent note class transition. 
For example, if note A is immediately followed by note B, 
note A is recorded as having a A-B transition. The notes
from each recording are treated separately.
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 0.2, 1.8), end = c(0.45, 0.7, 0.4, 2), pos = c(1,2,1,2),sound.files = c("JS001.wav","JS001.wav","JS002.wav","JS002.wav") , note_label = "Curve")
add_transitions(unit_table)
}
