% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transitions_dep.R
\name{add_transitions_dep}
\alias{add_transitions_dep}
\title{Subfunction for add_transitions}
\usage{
add_transitions_dep(table)
}
\arguments{
\item{table:}{Ordered unit table of notes from one song.}
}
\value{
A unit table with a transitions column.
}
\description{
Used in add_transitions() to add a transitions column for an ordered
unit table from one song. Designed only for use within add_transitions().
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 0.2, 1.8), end = c(0.45, 0.7, 0.4, 2), pos = c(1,2,3,4),sound.files ="JS001.wav" , note_label = "Curve")
add_transitions_dep(unit_table)
}
