% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap.R
\name{check_overlap}
\alias{check_overlap}
\title{check_overlap function}
\usage{
check_overlap(note1, note2)
}
\arguments{
\item{note1}{: A 2-dim numeric vector.}

\item{note2}{: A 2-dim numeric vector.}
}
\value{
A numeric scalar value indicating the amount of overlap. If
the intervals do not overlap, the value is 0.
}
\description{
Takes 2 intervals and computes how much they overlap.
}
\examples{
note1 = c(1,5), note2 = c(3,4)
check_overlap(note1, note2)
}
