% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_excision.R
\name{compare_excision}
\alias{compare_excision}
\title{compare_excision function}
\usage{
compare_excision(table1, table2)
}
\arguments{
\item{table1:}{A unit table tibble containing the original sound file, 
start/end times of the note and the note position.}

\item{table2:}{A unit table tibble containing the original sound file, 
start/end times of the note and the note position.}
}
\value{
Tibble containing the sound file, note positions in tables 1 and 2,
the square difference in the note intervals and a logical (matched) indicating
whether a matching note was found.
}
\description{
Takes two unit table of notes and compares the note excisions. 
Table1 corresponds to manually excised notes, Table2 corresponds
to computationally excised notes. The function goes through each 
note in Table1, tries to match it to a note in Table2 and 
quantifies the difference.
}
\examples{
table1 = tibble(start = c(0.35, 0.55), end = c(0.49, 0.7), sound.files = "JS001.wav", pos =c(2,3))
table2 = tibble(start = c(0.37, 0.6), end = c(0.45, 0.7), sound.files = "JS001.wav", pos =c(2,4))
}
