% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IOI.R
\name{compute_IOI}
\alias{compute_IOI}
\title{compute_IOI}
\usage{
compute_IOI(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
: A dataframe containing the ID, recording and IOIs of the start and end notes, note classes.
}
\description{
Takes a unit table and computes the inter-onset-intervals for a single sound file.
}
\examples{
 unit_table = tibble::tibble(start = c(0.37, 0.6, 1.2, 1.9), end = c(0.45, 0.7, 1.4, 2),
pos = seq(1,4),sound.files = "JS001.wav" , note_label = rep(c("A","B"),2))
compute_IOI(unit_table)
}
