% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_oscores.R
\name{compute_oscores}
\alias{compute_oscores}
\title{compute_oscores function}
\usage{
compute_oscores(manual_table, comp_table)
}
\arguments{
\item{manual_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}

\item{comp_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{

}
\description{
Compares a computational unit table against a manual unit table using the 
overlap scores defined in overlap_score. Function attempts to match 
each computational note to a note in the manual table with the most 
overlap.
}
\examples{
table1 = tibble::tibble(start = c(0.30, 0.55, 1.5, 2.5), end = c(0.51, 0.7, 2.2, 3), 
sound.files = "JS001.wav", pos =c(1,2,3,4), note_label = "Curve")
table2 = tibble::tibble(start = c(0.25, 0.65, 2), end = c(0.45, 0.75, 2.7), 
sound.files = "JS001.wav", pos =c(1,2,3), note_label = "Unclassifed")
compute_oscores(table1,table2)
}
