% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_oscores_dep.R
\name{compute_oscores_dep}
\alias{compute_oscores_dep}
\title{compute_oscores_dep}
\usage{
compute_oscores_dep(note, table)
}
\arguments{
\item{note}{: 2-dimensional numeric vector.}

\item{table}{:A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
A numeric vector containing the 3 overlap scores.
}
\description{
Dependency function for compute_oscores. Takes a note and matches it with
a note in a table with the most overlap. Computes the overlap scores between
the two notes.
}
\examples{
table1 = tibble::tibble(start = c(0.30, 0.55, 1.5, 2.5), end = c(0.51, 0.7, 2.2, 3), 
sound.files = "JS001.wav", pos =c(1,2,3,4), note_label = "Curve")
note = c(0.4,0.54)
}
