% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_score.R
\name{gap_score}
\alias{gap_score}
\title{gap_score function}
\usage{
gap_score(gap_table, pop_means)
}
\arguments{
\item{gap_table:}{A tibble as produced using the get_gaps function. It has Bird.ID, gap_dur,
selec and transitions.}

\item{pop_means:}{A tibble containing the mean gap lengths for each transition type. Columns should be
transitions and mean_gap.}
}
\value{

}
\description{
Computes the gap score for a bird, using the unit table derived from all its recordings.
}
\details{
For each transition type AB, compute

\deqn{score_AB = log( individual mean AB gaps / population mean AB gaps )}

The final score is the avaerage of the scores for every transition type in the individual bird.
}
\examples{
   unit_table = tibble::tibble(start = c(0.37, 0.6, 0.75, 1.2, 1.8, 2.5), 
end = c(0.45, 0.7, 0.9, 1.4, 2, 3), 
selec = seq(6),
sound.files = "JS001.wav", 
note_label = rep(c("A","B","C"),2))
gap_table = get_gaps(unit_table)
gap_score(gap_table)
}
