% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duration.R
\name{get_duration}
\alias{get_duration}
\title{get_duration function}
\usage{
get_duration(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
: A numeric scalar, the end of the first note minus the start of the first note.
}
\description{
Computes the duration of a song using its unit table.
}
\examples{
  unit_table = tibble::tibble(start = c(0.37, 0.6, 0.75, 0.2, 1.8, 2.5),
end = c(0.45, 0.7, 0.9, 0.4, 2, 3), 
sound.files = "JS001.wav",
select = seq(6),
note_label = "Curve")
get_duration(unit_table)
}
