% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tempo.R
\name{get_tempo}
\alias{get_tempo}
\title{get_tempo function}
\usage{
get_tempo(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
A 2 column dataframe containing each of the recording and its tempo.
}
\description{
Takes a unit table and computes the number of notes/sec (tempo) for
each recording.
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 0.2, 1.8), end = c(0.45, 0.7, 0.4, 2), pos = c(1,2,1,2),sound.files = c("JS001.wav","JS001.wav","JS002.wav","JS002.wav") , note_label = "Curve")
get_tempo(unit_table)
}
