% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_score_dep.R
\name{get_var_score_dep}
\alias{get_var_score_dep}
\title{get_var_score_dep function}
\usage{
get_var_score_dep(gap_table)
}
\arguments{
\item{gap_table}{}
}
\value{

}
\description{
Dependency function for get_var_score. Not to be used as a standalone. Takes a gap_table for gaps
belonging to the same transition type and computes the variability score. The formula is 
\eqn{score = var{ \frac{1}{\bar{G}_{ab} } } }.
}
