% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_score_dep2.R
\name{get_var_score_dep2}
\alias{get_var_score_dep2}
\title{get_var_score_dep2}
\usage{
get_var_score_dep2(transition, gap_table)
}
\arguments{
\item{transition:}{A string indicating the transition type.}

\item{gap_table:}{a gap table produced from the get_gaps function}
}
\value{
A numeric variable for the mean coefficient of variantion
}
\description{
Dependency function for gap_var_score. Not to be used alone. Computes the weighted 
population mean coefficient of variation for a transition type.
}
\examples{
  unit_table1 = tibble::tibble(start = c(0.37, 0.6, 0.75, 0.95, 1.8, 2.5),
end = c(0.45, 0.7, 0.9, 1.4, 2, 3),
selec = seq(6),
sound.files = c(rep("JS001.wav",6)),
note_label = c(rep("A",3), rep("B",3) ))
unit_table2 = tibble::tibble(start = c(0.30, 0.5, 0.75, 0.95, 1.8, 2.5),
                             end = c(0.45, 0.7, 0.9, 1.4, 2, 3),
                             selec = seq(6),
                             sound.files = c(rep("JS002.wav",6)),
                             note_label = c(rep("A",2), rep("B",4) ))
gap_table = get_gaps(rbind(unit_table1,unit_table2))

transition = "B-B"
output = get_var_score_dep2(transition, gap_table)
}
