% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_intro.R
\name{identify_intro}
\alias{identify_intro}
\title{identify_intro function}
\usage{
identify_intro(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
The input unit table with an additional column indicating whether 
each note is in an introduction or not.
}
\description{
Takes a unit table and identifies the notes that are part of the introduction. 
The first note is always part of the introduction. Subsequent notes are included
into the introduction until there are at least 5 continguous notes which are 
<0.25s apart.
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 0.75, 0.2, 1.8, 2.5), end = c(0.45, 0.7, 0.9, 0.4, 2, 3), pos = c(1,2,3,1,2,3),sound.files = c(rep("JS001.wav",3),rep("JS002.wav", 3)), note_label = "Curve")
identify_intro(intro_table)
}
