% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_intro_dep.R
\name{identify_intro_dep}
\alias{identify_intro_dep}
\title{identify_intro_dep}
\usage{
identify_intro_dep(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label. Notes are 
all from the same recording.}
}
\value{
A logical vector indicating
whether each note has been assigned to the introduction.
}
\description{
Dependency for identify_intro. Takes a unit table and assigns notes
to the introduction. The notes from the input are all from the same
recording.
}
\examples{
unit_table = tibble::tibble(start = c(0.2, 0.6, 0.75, 1.1, 1.8, 2.5, 4, 5, 6, 7, 8),
end = c(0.25, 0.7, 0.9, 1.6, 1.9, 3, 4.2, 5.5, 6.8, 7.6, 8.4), selec = seq(11),sound.files = "JS001.wav",
note_label = "Wave")
identify_intro_dep(unit_table)

}
