% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_compare.R
\name{note_compare}
\alias{note_compare}
\title{note_compare function}
\usage{
note_compare(note, unit_table)
}
\arguments{
\item{note}{: A unit table tibble containing the original sound file, 
start/end times of the note and the note class.}

\item{unit_table}{: A tibble of notes containing the original sound file, 
start/end times of the note and the note class.}
}
\value{
A tibble with the sound file, the square difference in the intervals, a 
logical indicating whether a note match was found and the note class. 
If no match it found, the difference is the square interval length of the input note.
}
\description{
Takes a note from one unit table and compares it with the corresponding note
found in another unit table.
}
\examples{
note = tibble(start = 0.35, end = 0.49, sound.files = "JS001.wav", pnote_label = "Curve")
unit_table = tibble(start = c(0.35, 0.55), end = c(0.49, 0.7), sound.files = "JS001.wav")
note_compare(note, unit_table)
}
