% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_match.R
\name{note_match}
\alias{note_match}
\title{note_match function}
\usage{
note_match(comp_table, manual_table)
}
\arguments{
\item{comp_table:}{A tibble of computationally excised notes containing the 
original sound file, start/end times of the note.}

\item{manual_table:}{A tibble of manually classified notes containing the 
original sound file, start/end times of the note and the note class.}
}
\value{
A vector of note classes for comp_table. The class is unclassified if
no matching note is found.
}
\description{
Takes a table of computationally excised notes and matches it to a table of
manually classified notes. Notes are considered matching if the midpoint 
of the computationally excised note is within the interval of a manually
classified note.
}
\examples{
comp_table = tibble(start = c(0.35, 0.55), end = c(0.49, 0.7), sound.files = "JS001.wav")
manual_table = tibble(start = c(0.37, 0.6), end = c(0.45, 0.7), sound.files = "JS001.wav", note_class = "Curve")
note_match(comp_table, manual_table)
}
