% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_scores.R
\name{overlap_scores}
\alias{overlap_scores}
\title{overlap_scores function}
\usage{
overlap_scores(table1, table2)
}
\arguments{
\item{table1}{}

\item{table2}{}
}
\value{

}
\description{
Compares two unit tables by computing the overlap scores for each pair
of matched notes. Notes are considered matching if their intervals overlap. 
3 scores are computed for each note pair. For any matched pair, denote
the intervals as the sets y and z. The scores are defined as follows.
\eqn{score1 = (y \cap z) / y},
\eqn{score2 = (y \cap z) / (y \cup z)},
\eqn{score3 = |(y \cap z)|^2 / |y| \times |z|}
}
\examples{
asdad
}
