% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_unit_table.R
\name{partition_unit_table}
\alias{partition_unit_table}
\title{partition_unit_table function}
\usage{
partition_unit_table(unit_table)
}
\arguments{
\item{unit_table:}{A unit table tibble containing the original sound file, 
start/end times of the note,the note position and note_label.}
}
\value{
A list of unit tables, one for each recording.
}
\description{
Separates a unit_table into a list of smaller unit_tables, 
one for each recording.
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 0.2, 1.8), end = c(0.45, 0.7, 0.4, 2), pos = c(1,2,1,2),sound.files = c("JS001.wav","JS001.wav","JS002.wav","JS002.wav") , note_label = "Curve")
partition_unit_table(unit_table)
}
