% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_n2.R
\name{scramble_n2}
\alias{scramble_n2}
\title{scramble n2 function}
\usage{
scramble_n2(ioi)
}
\arguments{
\item{ioi}{: An IOI table output by compute_IOI}
}
\value{
List of 2. 1:A permuted IOI table where each note2 is replaced by another note of the same class. 
2: Vector of row indices for the permuted IOI table
}
\description{
Takes the IOI table (output of compute_IOI) and permutes note2 (along with t2) according to
their note classes. I.e. Curve notes are replaced by other curve notes within the same song
}
\examples{
unit_table = tibble::tibble(start = c(0.37, 0.6, 1.2, 1.9), end = c(0.45, 0.7, 1.4, 2),
pos = seq(1,4),sound.files = "JS001.wav" , note_label = rep(c("A","B"),2))
ioi = compute_IOI(unit_table)
perm_ioi = scramble_n2(ioi)
}
