% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_within_groups.R
\name{scramble_within_groups}
\alias{scramble_within_groups}
\title{scramble_within_groups function}
\usage{
scramble_within_groups(data, group_index)
}
\arguments{
\item{data}{: A dataframe. One column is a categorical variable indicating group.}

\item{group_index}{: Column index of the variable indicating group.}
}
\value{
: A dataframe with permuted rows. Rows are only permuted within groups. Permutations are done via sampling with replacement.
}
\description{
Permutes the rows of a dataframe within groups.
}
\examples{
my_data <- data.frame(
Col1 = c(1, 2, 3, 4, 5, 6),
Col2 = c("A", "B", "A", "B", "C", "C"))
scramble_within_groups(my_data)
}
