% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{add_metadata function}
\usage{
add_metadata(data, metadata, ID)
}
\arguments{
\item{data}{: A dataframe with ID and some other columns}

\item{metadata:}{A metadata table with ID}

\item{ID:}{A character string indicating the ID column}
}
\value{
The original dataframe with the additional columns from the 
metadata.
}
\description{
Add the metadata to a dataframe based on a common ID column.
}
\examples{
data = tibble::tibble(x = c(1,2), ID = c("JS001", "JS002"))
metadata = tibble::tibble(y = c(3,4), ID = c("JS001", "JS002"))
}
